requirements.txt
# Deep Learning Frameworks
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Progress bars
tqdm>=4.65.0

# Testing
pytest>=7.3.0

# System monitoring
psutil>=5.9.0

# Quantization (optional - uncomment if needed)
# autoawq>=0.1.8
# auto-gptq>=0.5.0

# Jupyter (optional - uncomment if needed)
# jupyter>=1.0.0
# ipykernel>=6.20.0
# ipywidgets>=8.0.0
